refactor: split frozen diff snapshot modules
